+++ libs/cardano-ledger-core/src/Cardano/Ledger/Hashes.hs
@@ -200,7 +200,6 @@ newtype ScriptHash
     ( NFData
     , NoThunks
     , ToCBOR
-    , FromCBOR
     , EncCBOR
     , DecCBOR
     , ToJSON
@@ -227,8 +226,6 @@ newtype VRFVerKeyHash (r :: KeyRoleVRF) = VRFVerKeyHash
     ( NFData
     , NoThunks
     , Generic
-    , ToCBOR
-    , FromCBOR
     , EncCBOR
     , DecCBOR
     , ToJSONKey
@@ -337,8 +334,6 @@ newtype SafeHash i = SafeHash (Hash.Hash HASH i)
     , NoThunks
     , NFData
     , SafeToHash
-    , ToCBOR
-    , FromCBOR
     , EncCBOR
     , DecCBOR
     , ToJSON