+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 1.20.0.0
 
+* Remove `ToCBOR` and `FromCBOR` instances for `TxIx`, `CertIx`, `SlotNo32`, `Ptr`, `VRFVerKeyHash`, `SafeHash`, `VoidEraRule`, `Language`, `SLanguage`
+* Remove `FromCBOR` instances for `ScriptHash`, `ChainCode`, `PlutusBinary`, `BootstrapWitness`, `WitVKey`
+* Remove `ToCBOR` and `FromCBOR` instances for `PlutusWithContext`
+  - Add `EncCBOR` and `DecCBOR` instances as replacement
+* Remove `ToCBOR (TxOut era)` superclass constraint from `EraTxOut`
 * Deprecate `BHeaderView` in favour `*EraBlockHeader` typeclasses.
   - Move `isOverlaySlot` to `Cardano.Ledger.Slot`.
   - Remove `PerasCert`, `PerasKey` and `validatePerasCert` to `dijkstra`.